combine TestDriver and TestNode
There's a lot of overlap, so might as well put them together.
