+++ spec/spec_helper.rb
@@ -3,7 +3,7 @@
 require_relative "support/configuration"
 require_relative "support/coverage"
 require_relative "support/file_helpers"
-require_relative "support/test_driver"
+require_relative "support/test_node"
 require_relative "support/webmock"
 
 require_relative "../lib/feed_pub"
